web/janky-wonky-thing/chal/janky-wonky-thing/requirements.txt
argon2-cffi
fastapi[standard]
joserfc
